@@ -14,6 +14,6 @@ requirements = ['six',
                 'varz>=0.3.0']
 
 setup(packages=find_packages(exclude=['docs']),
-      python_requires='>=3.5',
+      python_requires='>=3.6',
       install_requires=requirements,
       include_package_data=True)
