@@ -1,10 +1,7 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import absolute_import, division, print_function
+import lab as B
+from stheno import Normal
 
 from wbml.vi import elbo
-from stheno import Normal
-import lab as B
 from .util import allclose
 
 
