@@ -1,9 +1,5 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import absolute_import, division, print_function
-
-from .util import *
-from .net import *
 from .data import *
-from .vi import *
 from .lmm import *
+from .net import *
+from .util import *
+from .vi import *
